refactor(dashboard): extract project mutations into a hook
- useProjectMutations owns save/delete/source writes with their cache
  invalidation and toasts; settings page drops ~70 lines of plumbing
